refactor: move primary-action-dropdown
