+++ packages/components/primary-action-dropdown/index.js
@@ -1,6 +1,3 @@
-export {
-  default,
-  Option,
-} from '../../../src/components/dropdowns/primary-action-dropdown';
+export { default, Option } from './src';
 
 export { default as version } from './version';
